fixes again

+++ modules/task_3/bakaeva_m_integrals_rectangles_method/integrals_rectangles_method.cpp
@@ -46,14 +46,14 @@ double getSequentialIntegrals(const int n, vector<pair<double, double> > a_b, do
 }
 
 double getParallelIntegrals(const int n, vector<pair<double, double> > a_b, double (*F)(vector<double>)) {
-    int size, rank;
+    size_t size, rank;
 
     //Число задействованных процессов
     MPI_Comm_size(MPI_COMM_WORLD, &size);
     //Получение номера текущего процесса в рамках коммуникатора
     MPI_Comm_rank(MPI_COMM_WORLD, &rank);
 
-    int countIntegrals = a_b.size();
+    size_t countIntegrals = a_b.size();
     vector<double> h(n);
     vector<pair<double, double>> ab(countIntegrals);
     size_t countElements;  // Количество всех одночленов
@@ -65,7 +65,7 @@ double getParallelIntegrals(const int n, vector<pair<double, double> > a_b, doub
             h[i] = (a_b[i].second - a_b[i].first) / n;
             ab[i] = a_b[i];
         }
-        int j = 0;
+        size_t j = 0;
         while (j != countIntegrals - 1) {
             countElements *= n;
             j++;
@@ -75,7 +75,7 @@ double getParallelIntegrals(const int n, vector<pair<double, double> > a_b, doub
     // Рассылаем данные всем процессам
     int length = n;  // Количество отрезков интегрирования
     MPI_Bcast(&countElements, 1, MPI_UNSIGNED, 0, MPI_COMM_WORLD);
-    MPI_Bcast(&length, countIntegrals, MPI_INT, 0, MPI_COMM_WORLD);
+    MPI_Bcast(&length, countIntegrals, MPI_UNSIGNED, 0, MPI_COMM_WORLD);
     MPI_Bcast(&h[0], countIntegrals, MPI_DOUBLE, 0, MPI_COMM_WORLD);
     MPI_Bcast(&ab[0], 2 * countIntegrals, MPI_DOUBLE, 0, MPI_COMM_WORLD);
 
@@ -94,16 +94,16 @@ double getParallelIntegrals(const int n, vector<pair<double, double> > a_b, doub
 
     // Каждый процесс вычисляет свое количество sum(f(x..))
     vector<vector<double>> forCalculateF(delta);
-    for (int i = 0; i < delta; i++) {
+    for (size_t i = 0; i < delta; i++) {
         int number = tmp + i;
-        for (int j = 0; j < countIntegrals - 1; j++) {
+        for (size_t j = 0; j < countIntegrals - 1; j++) {
             forCalculateF[i].push_back(ab[j].first + h[j] * (number % n) + h[j] * 0.5);
         }
     }
 
     double result = 0.0;
-    for (int i = 0; i < delta; i++) {
-        for (int j = 0; j < n; j++) {
+    for (size_t i = 0; i < delta; i++) {
+        for (size_t j = 0; j < n; j++) {
             forCalculateF[i].push_back(ab[countIntegrals - 1].first + (j + 0.5) * h[countIntegrals - 1]);
             result += F(forCalculateF[i]);
             forCalculateF[i].pop_back();
@@ -114,7 +114,7 @@ double getParallelIntegrals(const int n, vector<pair<double, double> > a_b, doub
     double Integral = 0.0;
     MPI_Reduce(&result, &Integral, 1, MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
     // Умножение на h по формуле
-    for (int i = 0; i < countIntegrals; i++) {
+    for (size_t i = 0; i < countIntegrals; i++) {
         Integral *= h[i];
     }
